requirements.txt
openai
langchain
docx2txt
pypdf
streamlit
chromadb
tiktoken
python-dotenv
wikipedia
streamlit-chat